update code from code review request
