+++ src/components/Alteration/index.ts
@@ -1,2 +1,3 @@
 export * from './Articles'
 export { default as AlterationSummary } from './AlterationSummary.vue'
+export { default as BcRegEntityDetails } from './BcRegEntityDetails.vue'
