+++ src/interfaces/resource-interfaces/entity-type-option-interface.ts
@@ -1,0 +1,5 @@
+export interface EntityTypeOption {
+    value: string;
+    shortDesc: string;
+    text: string;
+}
