+++ src/interfaces/resource-interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from './component-resource-interfaces'
 export * from './resource-interface'
 export * from './special-resolution-sample-form-interface'
+export * from './entity-type-option-interface'
