+++ src/resources/Alteration/BEN.ts
@@ -20,17 +20,17 @@ export const BenAlterationResource: ResourceIF = {
     entityTypeOptions: [
       {
         value: 'BEN',
-        SHORT_DESC: 'BC Benefit Company',
+        shortDesc: 'BC Benefit Company',
         text: 'BC Benefit Company'
       },
       {
         value: 'BC',
-        SHORT_DESC: 'BC Limited Company',
+        shortDesc: 'BC Limited Company',
         text: 'BC Limited Company'
       },
       {
         value: 'CC',
-        SHORT_DESC: 'BC Community Contribution Company',
+        shortDesc: 'BC Community Contribution Company',
         text: 'BC Community Contribution Company'
       }
     ],
